backend/requirements.txt
fastapi[standard]==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
redis==5.2.1
httpx==0.28.1
boto3==1.35.76
dramatiq[redis]==1.17.1
pytest==8.3.4
pytest-asyncio==0.24.0
ruff==0.8.4
